update example codes.
