+++ zio-http-example/src/main/scala/example/auth/bearer/jwt/refresh/core/User.scala
@@ -7,7 +7,7 @@ case class User(username: String, password: Secret, email: String, roles: Set[St
 
 sealed trait UserServiceError
 object UserServiceError {
-  case class UserNotFound(username: String) extends UserServiceError
+  case class UserNotFound(username: String)      extends UserServiceError
   case class UserAlreadyExists(username: String) extends UserServiceError
 }
 
@@ -39,14 +39,14 @@ case class UserServiceLive(users: Ref[Map[String, User]]) extends UserService {
 
   def updateEmail(username: String, newEmail: String): IO[UserNotFound, Unit] = for {
     currentUsers <- users.get
-    user <- ZIO.fromOption(currentUsers.get(username)).orElseFail(UserNotFound(username))
-    _ <- users.update(_.updated(username, user.copy(email = newEmail)))
+    user         <- ZIO.fromOption(currentUsers.get(username)).orElseFail(UserNotFound(username))
+    _            <- users.update(_.updated(username, user.copy(email = newEmail)))
   } yield ()
 
   def validateCredentials(username: String, password: String): IO[UserNotFound, User] =
     for {
       user <- getUser(username)
-      _ <- ZIO.when(user.password != Secret(password)) {
+      _    <- ZIO.when(user.password != Secret(password)) {
         ZIO.fail(UserNotFound(username)) // Using UserNotFound to indicate invalid credentials
       }
     } yield user
@@ -54,9 +54,9 @@ case class UserServiceLive(users: Ref[Map[String, User]]) extends UserService {
 
 object UserService {
   private val initialUsers = Map(
-    "john" -> User("john", Secret("password123"), "john@example.com", Set("user")),
-    "jane" -> User("jane", Secret("secret456"), "jane@example.com", Set("user")),
-    "admin" -> User("admin", Secret("admin123"), "admin@company.com", Set("user", "admin"))
+    "john"  -> User("john", Secret("password123"), "john@example.com", Set("user")),
+    "jane"  -> User("jane", Secret("secret456"), "jane@example.com", Set("user")),
+    "admin" -> User("admin", Secret("admin123"), "admin@company.com", Set("user", "admin")),
   )
 
   val live: ZLayer[Any, Nothing, UserService] =
