+++ zio-http-example/src/main/scala/example/auth/bearer/jwt/refresh/core/UserInfo.scala
@@ -3,13 +3,12 @@ package example.auth.bearer.jwt.refresh.core
 import zio.json._
 
 case class UserInfo(
-                     @jsonField("sub")
-                     username: String,
-                     email: String,
-                     roles: Set[String]
-                   )
+  @jsonField("sub")
+  username: String,
+  email: String,
+  roles: Set[String],
+)
 
 object UserInfo {
   implicit val codec: JsonCodec[UserInfo] = DeriveJsonCodec.gen
 }
-
